fetcher_service/requirements.txt
flask
nordpool
python-dateutil
apscheduler
requests